extending write_ascii_output (*.asc) files with sub-options ascii_out_4326 (reprojecting to lat/lon grid), ascii_out_json (using a JSON variant of AAIGRID) and ascii_out_uv (writing u,v wind field grids instead of ang,vel). All of them are off by default. There also are options that represent the old behavior, which are set by default: ascii_out_utm for UTM grids and ascii_out_aaigrid for AAIGRID text format.
Filenames are derived from the configured ascii grid output filenames

All flags are orthogonal, i.e. all respective output can be generated in a single run

set minimum version in top CMakeLists to 3 to avoid deprecated warnings

link libninja static on APPLE so that we only have 3rd party dylib dependencies that are OS/user-wide installed. This allows to run WindNinja_cli directly from the build dir
without having to fiddle with LD_LIBRARY_PATH / DYLD_LIBRARY_PATH
