+++ src/ninja/WindNinjaInputs.cpp
@@ -73,9 +73,14 @@ WindNinjaInputs::WindNinjaInputs()
     wxModelGoogSpeedScaling = KmlVector::equal_interval;
     wxModelGoogLineWidth = 3.0;
     shpOutFlag = false;
+    
     asciiOutFlag = false;
+    asciiAaigridOutFlag = true;
+    asciiJsonOutFlag = false;
     ascii4326OutFlag = false;
-    json4326OutFlag = false;
+    asciiUtmOutFlag = true;
+    asciiUvOutFlag = false;
+    
     wxModelShpOutFlag = false;
     wxModelAsciiOutFlag = false;
     txtOutFlag = false;
@@ -175,159 +180,7 @@ WindNinjaInputs::~WindNinjaInputs()
 WindNinjaInputs::WindNinjaInputs(const WindNinjaInputs &rhs)
 : ninjaTime(boost::local_time::not_a_date_time)
 {
-  armySize = rhs.armySize;
-  hSpdMemDs = rhs.hSpdMemDs;
-  hDirMemDs = rhs.hDirMemDs;
-  hDustMemDs = rhs.hDustMemDs;
-  
-  vegetation = rhs.vegetation;
-
-  initializationMethod = rhs.initializationMethod;
-  forecastFilename = rhs.forecastFilename;
-  inputSpeedUnits = rhs.inputSpeedUnits;
-  outputSpeedUnits = rhs.outputSpeedUnits;
-  inputSpeed = rhs.inputSpeed;
-  inputDirection = rhs.inputDirection;
-  inputWindHeightUnits = rhs.inputWindHeightUnits;
-  inputWindHeight = rhs.inputWindHeight;
-  outputWindHeightUnits = rhs.outputWindHeightUnits;
-  outputWindHeight = rhs.outputWindHeight;
-  
-  realStations=rhs.realStations;
-
-  stations = rhs.stations;
-  wxStationFilename = rhs.wxStationFilename;
-  stationsScratch = rhs.stationsScratch;
-  stationsOldInput = rhs.stationsOldInput;
-  stationsOldOutput = rhs.stationsOldOutput;
-  stationFetch=rhs.stationFetch;
-  matchWxStations = rhs.matchWxStations;
-  outer_relax = rhs.outer_relax;
-  CPLDebug("NINJA", "Setting NINJA_POINT_MATCH_OUT_RELAX to %lf", outer_relax);
-  diurnalWinds = rhs.diurnalWinds;
-#ifdef NINJAFOAM
-    nIterations = rhs.nIterations;
-    meshCount = rhs.meshCount;
-    ninjafoamMeshChoice = rhs.ninjafoamMeshChoice;
-    stlFile = rhs.stlFile;
-    foamVelocityGrid = rhs.foamVelocityGrid;
-    foamAngleGrid = rhs.foamAngleGrid;
-#endif
-  outputPointsFilename = rhs.outputPointsFilename;
-  inputPointsFilename = rhs.inputPointsFilename;
-  pointsNamesList = rhs.pointsNamesList;
-  latList = rhs.latList;
-  lonList = rhs.lonList;
-  heightList = rhs.heightList;
-
-  ninjaTime = rhs.ninjaTime;
-  if(rhs.ninjaTimeZone.get() == NULL)   //If pointer is NULL
-      ninjaTimeZone.reset();
-  else if(rhs.ninjaTimeZone->to_posix_string().empty()) //If pointer is good, but posix string is empty
-      ninjaTimeZone.reset();
-  else  //Else it is a good time zone pointer, so set accordingly
-      ninjaTimeZone = boost::local_time::time_zone_ptr (new boost::local_time::posix_time_zone(rhs.ninjaTimeZone->to_posix_string()));
-
-  airTemp = rhs.airTemp;
-  cloudCover = rhs.cloudCover;
-  latitude = rhs.latitude;
-  longitude = rhs.longitude;
-  numberCPUs = rhs.numberCPUs;
-  outputBufferClipping = rhs.outputBufferClipping;
-  writeAtmFile = rhs.writeAtmFile;
-  googOutFlag = rhs.googOutFlag;
-  googSpeedScaling = rhs.googSpeedScaling;
-  googLineWidth = rhs.googLineWidth;
-  googColor = rhs.googColor;
-  googVectorScale = rhs.googVectorScale;
-  wxModelGoogOutFlag = rhs.wxModelGoogOutFlag;
-  wxModelGoogSpeedScaling = rhs.wxModelGoogSpeedScaling;
-  wxModelGoogLineWidth = rhs.wxModelGoogLineWidth;
-  shpOutFlag = rhs.shpOutFlag;
-  pdfOutFlag = rhs.pdfOutFlag;
-  pdfDEMFileName = rhs.pdfDEMFileName;
-  pdfResolution = rhs.pdfResolution;
-  pdfLineWidth  = rhs.pdfLineWidth ;
-  pdfUnits = rhs.pdfUnits;
-  pdfFile = rhs.pdfFile;
-  pdfBaseType = rhs.pdfBaseType;
-  pdfWidth = rhs.pdfWidth;
-  pdfHeight = rhs.pdfHeight;
-  pdfDPI = rhs.pdfDPI;
-  asciiOutFlag = rhs.asciiOutFlag;
-  ascii4326OutFlag = rhs.ascii4326OutFlag;
-  json4326OutFlag = rhs.json4326OutFlag;
-  wxModelShpOutFlag = rhs.wxModelShpOutFlag;
-  wxModelAsciiOutFlag = rhs.wxModelAsciiOutFlag;
-  txtOutFlag = rhs.txtOutFlag;
-  volVTKOutFlag = rhs.volVTKOutFlag;
-  kmlFile = rhs.kmlFile;
-  kmzFile = rhs.kmzFile;
-  wxModelKmlFile = rhs.wxModelKmlFile;
-  wxModelKmzFile = rhs.wxModelKmzFile;
-  kmzResolution = rhs.kmzResolution;
-  kmzUnits = rhs.kmzUnits;
-  shpFile = rhs.shpFile;
-  dbfFile = rhs.dbfFile;
-  wxModelShpFile = rhs.wxModelShpFile;
-  wxModelDbfFile = rhs.wxModelDbfFile;
-  shpResolution = rhs.shpResolution;
-  shpUnits = rhs.shpUnits;
-
-
-  cldFile = rhs.cldFile;
-  velFile = rhs.velFile;
-  wxModelCldFile = rhs.wxModelCldFile;
-  wxModelVelFile = rhs.wxModelVelFile;
-  velResolution = rhs.velResolution;
-  velOutputFileDistanceUnits = rhs.velOutputFileDistanceUnits;
-  angFile = rhs.angFile;
-  atmFile = rhs.atmFile;
-  wxModelAngFile = rhs.wxModelAngFile;
-  angResolution = rhs.angResolution;
-  angOutputFileDistanceUnits = rhs.angOutputFileDistanceUnits;
-  legFile = rhs.legFile;
-  dateTimeLegFile = rhs.dateTimeLegFile;
-  volVTKFile = rhs.volVTKFile;
-  keepOutGridsInMemory = rhs.keepOutGridsInMemory;
-  customOutputPath = rhs.customOutputPath;
-  
-#ifdef NINJA_SPEED_TESTING
-  speedDampeningRatio = rhs.speedDampeningRatio;
-#endif
-  
-  downDragCoeff = rhs.downDragCoeff;
-  downEntrainmentCoeff = rhs.downDragCoeff;
-  upDragCoeff = rhs.upDragCoeff;
-  upEntrainmentCoeff = rhs.upEntrainmentCoeff;
-#ifdef FRICTION_VELOCITY
-  frictionVelocityFlag = rhs.frictionVelocityFlag; 
-  frictionVelocityCalculationMethod = rhs.frictionVelocityCalculationMethod;
-  ustarFile = rhs.ustarFile;
-#endif
-#ifdef EMISSIONS
-  dustFlag = rhs.dustFlag;
-  dustFilename = rhs.dustFilename;
-  dustFileOut = rhs.dustFileOut;
-  geotiffOutFlag = rhs.geotiffOutFlag;
-  dustFile = rhs.dustFile;
-  ustarFile = rhs.ustarFile;
-#endif
-  
-  stabilityFlag = rhs.stabilityFlag;
-  alphaStability = rhs.alphaStability;
-  
-  outputPath = rhs.outputPath;
-
-  //class crap
-  dem = rhs.dem;
-  surface = rhs.surface;
-
-#ifdef _OPENMP
-  omp_set_nested(false);
-  omp_set_dynamic(false);
-#endif //_OPENMP
-
+  *this = rhs;
 }
 
 /**
@@ -442,9 +295,14 @@ WindNinjaInputs &WindNinjaInputs::operator=(const WindNinjaInputs &rhs)
       wxModelGoogSpeedScaling = rhs.wxModelGoogSpeedScaling;
       wxModelGoogLineWidth = rhs.wxModelGoogLineWidth;
       shpOutFlag = rhs.shpOutFlag;
+
       asciiOutFlag = rhs.asciiOutFlag;
+      asciiAaigridOutFlag = rhs.asciiAaigridOutFlag;
+      asciiJsonOutFlag = rhs.asciiJsonOutFlag;
+      asciiUtmOutFlag = rhs.asciiUtmOutFlag;
       ascii4326OutFlag = rhs.ascii4326OutFlag;
-      json4326OutFlag = rhs.json4326OutFlag;
+      asciiUvOutFlag = rhs.asciiUvOutFlag;
+      
       wxModelShpOutFlag = rhs.wxModelShpOutFlag;
       wxModelAsciiOutFlag = rhs.wxModelAsciiOutFlag;
       txtOutFlag = rhs.txtOutFlag;
