+++ src/ninja/cli.cpp
@@ -91,6 +91,48 @@ pair<string, string> at_option_parser(string const&s)
         return pair<string, string>();
 }
 
+// if we have an 'elevation_file' program option check if file exists and has a non-geographic srs.
+// if the srs is geographic, try to convert to a UTM file in the configured 'output_path' (or current dir if not set)
+// return address of a string that points to a valid non-geographic file or NULL if none was found or could be constructed
+const std::string* get_checked_elevation_file (po::variables_map& vm)
+{
+    if (vm.count("elevation_file")) {
+        const string* filename = &vm["elevation_file"].as<string>();
+        if (!CPLCheckForFile( (char*)filename->c_str(), NULL)){
+            throw std::logic_error( string("elevation_file " + *filename + " not found"));
+        }
+
+        GDALDataset *pSrcDS = (GDALDataset*) GDALOpen(filename->c_str(), GA_ReadOnly);
+        if (pSrcDS) {
+            const OGRSpatialReference *pSR = pSrcDS->GetSpatialRef();
+            if (pSR && pSR->IsGeographic()){
+                cout << "provided elevation_file " << *filename << " is geographic, converting..\n";
+                string output_path = vm.count("output_path") ? vm["output_path"].as<string>().c_str() : "";
+                string new_filename = derived_pathname( filename->c_str(), output_path.c_str(), "\\.([^.]+)$", "-utm.$1");
+                GDALDataset *pDstDS = gdalWarpToUtm( new_filename.c_str(), pSrcDS);
+                if (pDstDS) {
+                    cout << "using warped UTM elevation_file " << new_filename << "\n";
+                    GDALClose(pDstDS);
+                    GDALClose(pSrcDS);
+                    return new string(new_filename);
+
+                } else { // converting to UTM elevation file failed
+                    GDALClose(pSrcDS);
+                    throw std::logic_error( string("elevation_file ") + *filename + " cannot be converted to UTM");
+                }
+
+            } else { // original elevation_file SRS is not geographic, use as-is
+                GDALClose(pSrcDS);
+                return filename;
+            }
+        } else { // GDALOpen of original elevation_file failed
+            throw std::logic_error( string("invalid elevation_file ") + *filename);
+        }
+    } else { // no elevation_file specified
+        return NULL; 
+    }
+}
+
 /**
  * Command line implementation (CLI) of WindNinja.  Can be run using command line args or
  * from an input file.
@@ -250,8 +292,13 @@ int windNinjaCLI(int argc, char* argv[])
                 ("units_shape_out_resolution", po::value<std::string>()->default_value("m"), "units of shapefile resolution (ft, m)")
                 ("write_wx_model_ascii_output", po::value<bool>()->default_value(false), "write ascii fire behavior output files for the raw wx model forecast (true, false)")
                 ("write_ascii_output", po::value<bool>()->default_value(false), "write ascii fire behavior output files (true, false)")
-                ("write_ascii_4326_output", po::value<bool>()->default_value(false), "write AAIGRID (ascii) fire behavior output files in EPSG:4326 lat/lon projection (true, false)")
-                ("write_json_4326_output", po::value<bool>()->default_value(false), "write AAIGRID (JSON) fire behavior output files in EPSG:4326 lat/lon projection (true, false)")
+
+                ("ascii_out_aaigrid", po::value<bool>()->default_value(true), "write ascii output as AAIGRID files (default: true, false)")
+                ("ascii_out_json", po::value<bool>()->default_value(false), "write ascii output as JSON files (true, default:false)")
+                ("ascii_out_4326", po::value<bool>()->default_value(false), "write ascii files as EPSG:4326 lat/lon grids (true, default:false)")
+                ("ascii_out_utm", po::value<bool>()->default_value(true), "write ascii files as UTM northing/easting grids (default: true, false)")
+                ("ascii_out_uv", po::value<bool>()->default_value(false), "write ascii files as u,v wind vector components (true, default:false)")
+
                 ("ascii_out_resolution", po::value<double>()->default_value(-1.0), "resolution of ascii fire behavior output files (-1 to use mesh resolution)")
                 ("units_ascii_out_resolution", po::value<std::string>()->default_value("m"), "units of ascii fire behavior output file resolution (ft, m)")
                 ("write_vtk_output", po::value<bool>()->default_value(false), "write VTK output file (true, false)")
@@ -548,6 +595,9 @@ int windNinjaCLI(int argc, char* argv[])
             }
         }
 
+        const std::string* elevation_file = get_checked_elevation_file(vm); // might either be NULL or set dynamically
+        std::string output_path = vm.count("output_path") ? vm["output_path"].as<std::string>() : "";
+
 #ifdef NINJAFOAM
         ninjaArmy windsim(1, vm["momentum_flag"].as<bool>()); //-Moved to header file
 #else
@@ -562,7 +612,7 @@ int windNinjaCLI(int argc, char* argv[])
         /* if elevation_file wasn't specified       */ 
         /*------------------------------------------*/            
          
-        if(vm["compute_emissions"].as<bool>() && !vm.count("elevation_file")){
+        if(vm["compute_emissions"].as<bool>() && !elevation_file){
             OGRDataSourceH hDS = 0;
             hDS = OGROpen(vm["fire_perimeter_file"].as<std::string>().c_str(), FALSE, 0);
             if (hDS == 0){
@@ -651,8 +701,7 @@ int windNinjaCLI(int argc, char* argv[])
                 }
             }
             
-            vm.insert(std::make_pair("elevation_file", po::variable_value(new_elev, false)));
-            po::notify(vm);
+            elevation_file = new string(new_elev);
         }
         #endif //EMISSIONS
 
@@ -763,8 +812,8 @@ int windNinjaCLI(int argc, char* argv[])
                 VSIUnlink(new_elev.c_str());
                 exit(1);
             }
-            vm.insert(std::make_pair("elevation_file", po::variable_value(vm["fetch_elevation"])));
-            po::notify(vm);
+
+            elevation_file = &vm["fetch_elevation"].as<std::string>();
             //std::cout << "Elevation file download complete." << std::endl;
         }
 
@@ -781,7 +830,7 @@ int windNinjaCLI(int argc, char* argv[])
         if( bFillNoData )
         {
             GDALDataset *poDS;
-            poDS = (GDALDataset*)GDALOpen(vm["elevation_file"].as<std::string>().c_str(), GA_Update);
+            poDS = (GDALDataset*)GDALOpen(elevation_file->c_str(), GA_Update);
             if(poDS == NULL)
             {
                 throw std::runtime_error("Could not open DEM for reading");
@@ -809,7 +858,7 @@ int windNinjaCLI(int argc, char* argv[])
             {
                 double longitude = 0;
                 double latitude = 0;
-                GDALDataset *poDS = (GDALDataset*)GDALOpen(vm["elevation_file"].as<std::string>().c_str(), GA_ReadOnly);
+                GDALDataset *poDS = (GDALDataset*)GDALOpen(elevation_file->c_str(), GA_ReadOnly);
                 if(poDS == NULL)
                 {
                     GDALClose((GDALDatasetH)poDS);
@@ -879,8 +928,7 @@ int windNinjaCLI(int argc, char* argv[])
                 try
                 {
                     model = wxModelInitializationFactory::makeWxInitializationFromId( model_type );
-                    std::string forecastFileName = model->fetchForecast( vm["elevation_file"].as<std::string>(),
-                                                                                vm["forecast_duration"].as<int>() );
+                    std::string forecastFileName = model->fetchForecast( *elevation_file, vm["forecast_duration"].as<int>() );
                     if(vm.count("start_year"))
                     {
                         conflicting_options(vm, "forecast_time", "start_year");
@@ -947,7 +995,7 @@ int windNinjaCLI(int argc, char* argv[])
                                       timeList,
                                       vm["momentum_flag"].as<bool>() );
 #else
-                    windsim.makeArmy( model->fetchForecast( vm["elevation_file"].as<std::string>(),
+                    windsim.makeArmy( model->fetchForecast( *elevation_file,
                                                             vm["forecast_duration"].as<int>() ),
                                                             osTimeZone,
                                                             timeList,
@@ -997,7 +1045,7 @@ int windNinjaCLI(int argc, char* argv[])
             if(vm["fetch_station"].as<bool>() == true) //download station and make appropriate size ninjaArmy
             {
                 const char *api_key_conf_opt = CPLGetConfigOption("CUSTOM_API_KEY","FALSE");
-                if(api_key_conf_opt!="FALSE")
+                if (strcmp(api_key_conf_opt,"FALSE") != 0)
                 {
                     std::ostringstream api_stream;
                     api_stream<<api_key_conf_opt;
@@ -1049,24 +1097,18 @@ int windNinjaCLI(int argc, char* argv[])
                     boost::posix_time::ptime noTime;
                     timeList.push_back(noTime);
                 }
+
                 //Generate a directory to store downloaded station data...
                 CPLDebug("STATION_FETCH","Generating Directory for Weather Stations");
-                if(vm.count("output_path")){
-                    stationPathName=pointInitialization::generatePointDirectory(vm["elevation_file"].as<std::string>(),
-                                                                                vm["output_path"].as<std::string>(),
+                stationPathName=pointInitialization::generatePointDirectory(*elevation_file,
+                                                                                output_path,
                                                                                 vm["fetch_current_station_data"].as<bool>());
-                }
-                else{ //if the user doesn't specify an output path
-                    stationPathName=pointInitialization::generatePointDirectory(vm["elevation_file"].as<std::string>(),
-                                                                                "",
-                                                                                vm["fetch_current_station_data"].as<bool>());
-                }
 //                stationPathName="blank";
                 pointInitialization::SetRawStationFilename(stationPathName); //Set this for fetching
                 //so that the fetchStationData function knows where to save the data
                 if (vm["fetch_type"].as<std::string>()=="bbox") //Get data from Bounding Box
                 {
-                    bool fetchSuccess = pointInitialization::fetchStationFromBbox(vm["elevation_file"].as<std::string>(),
+                    bool fetchSuccess = pointInitialization::fetchStationFromBbox(*elevation_file,
                                                             timeList, osTimeZone,
                                                             vm["fetch_current_station_data"].as<bool>());
                     if(fetchSuccess==false) //If we fail to download any data
@@ -1076,7 +1118,7 @@ int windNinjaCLI(int argc, char* argv[])
                     }
 
                     //                    pointInitialization::writeStationLocationFile(vm["elevation_file"].as<std::string>());
-                    pointInitialization::writeStationLocationFile(stationPathName,vm["elevation_file"].as<std::string>(),vm["fetch_current_station_data"].as<bool>());
+                    pointInitialization::writeStationLocationFile(stationPathName,*elevation_file,vm["fetch_current_station_data"].as<bool>());
                     
                 }
                 else if (vm["fetch_type"].as<std::string>()=="stid")
@@ -1091,8 +1133,8 @@ int windNinjaCLI(int argc, char* argv[])
                         pointInitialization::removeBadDirectory(stationPathName); //delete the generated dir
                         throw std::runtime_error(pointInitialization::error_msg);
                     }
-//                    pointInitialization::writeStationLocationFile(vm["elevation_file"].as<std::string>()); 
-                    pointInitialization::writeStationLocationFile(stationPathName,vm["elevation_file"].as<std::string>(),vm["fetch_current_station_data"].as<bool>());
+//                    pointInitialization::writeStationLocationFile(*elevation_file); 
+                    pointInitialization::writeStationLocationFile(stationPathName,*elevation_file,vm["fetch_current_station_data"].as<bool>());
                     
                 }
                 else //If something else bad happens
@@ -1105,13 +1147,13 @@ int windNinjaCLI(int argc, char* argv[])
                 windsim.makeStationArmy(timeList,
                                         osTimeZone,
                                         stationPathName,
-                                        vm["elevation_file"].as<std::string>(),
+                                        *elevation_file,
                                         vm["match_points"].as<bool>(),false);
 
                 if(vm["fetch_metadata"].as<bool>() == true) //fetches metadata
                 {
                     pointInitialization::fetchMetaData(vm["metadata_filename"].as<std::string>(),
-                            vm["elevation_file"].as<std::string>(),true);
+                            *elevation_file,true);
                 }
             }
             else if (vm["fetch_station"].as<bool>() == false) //If we aren't fetching, look for on disk files
@@ -1180,7 +1222,7 @@ int windNinjaCLI(int argc, char* argv[])
                         sFiles.push_back(vm["wx_station_filename"].as<std::string>());
                         pointInitialization::storeFileNames(sFiles);
                         windsim.makeStationArmy(timeList,osTimeZone,vm["wx_station_filename"].as<std::string>(),
-                                vm["elevation_file"].as<std::string>(),vm["match_points"].as<bool>(),false);
+                                *elevation_file,vm["match_points"].as<bool>(),false);
                     }
                     if(fileSubFormat==1) //not a time series
                     {
@@ -1191,7 +1233,7 @@ int windNinjaCLI(int argc, char* argv[])
                         sFiles.push_back(vm["wx_station_filename"].as<std::string>());
                         pointInitialization::storeFileNames(sFiles);
                         windsim.makeStationArmy(timeList,osTimeZone,vm["wx_station_filename"].as<std::string>(),
-                                vm["elevation_file"].as<std::string>(),vm["match_points"].as<bool>(),false);
+                                *elevation_file,vm["match_points"].as<bool>(),false);
                     }
                 }
                 else if (stationFormat==1) //old format
@@ -1200,7 +1242,7 @@ int windNinjaCLI(int argc, char* argv[])
                     boost::posix_time::ptime noTime;
                     timeList.push_back(noTime);
                     windsim.makeStationArmy(timeList,osTimeZone,vm["wx_station_filename"].as<std::string>(),
-                            vm["elevation_file"].as<std::string>(),vm["match_points"].as<bool>(),false);
+                            *elevation_file,vm["match_points"].as<bool>(),false);
                 }
                 else if (stationFormat==3) // New Format where there are multiple station files
                 {
@@ -1233,7 +1275,7 @@ int windNinjaCLI(int argc, char* argv[])
                     sFiles=pointInitialization::openCSVList(vm["wx_station_filename"].as<std::string>());                   
                     pointInitialization::storeFileNames(sFiles);
                     windsim.makeStationArmy(timeList,osTimeZone,vm["wx_station_filename"].as<std::string>(),
-                            vm["elevation_file"].as<std::string>(),vm["match_points"].as<bool>(),false);
+                            *elevation_file,vm["match_points"].as<bool>(),false);
                 }
                 else if (stationFormat==4) // New Format where there are multiple one step recent station files
                 {
@@ -1247,7 +1289,7 @@ int windNinjaCLI(int argc, char* argv[])
                     sFiles=pointInitialization::openCSVList(vm["wx_station_filename"].as<std::string>());
                     pointInitialization::storeFileNames(sFiles);
                     windsim.makeStationArmy(timeList,osTimeZone,vm["wx_station_filename"].as<std::string>(),
-                            vm["elevation_file"].as<std::string>(),vm["match_points"].as<bool>(),false);
+                            *elevation_file,vm["match_points"].as<bool>(),false);
                 }
                 else
                 {
@@ -1274,9 +1316,9 @@ int windNinjaCLI(int argc, char* argv[])
 
             windsim.setNumberCPUs( i_, vm["num_threads"].as<int>() );
 
-            //windsim.ninjas[i_].readInputFile(vm["elevation_file"].as<std::string>());
+            //windsim.ninjas[i_].readInputFile(*elevation_file);
             
-            windsim.setDEM( i_, vm["elevation_file"].as<std::string>() );
+            windsim.setDEM( i_, *elevation_file );
             windsim.setPosition( i_ );    //get position from DEM file
             
             #ifdef NINJAFOAM
@@ -1473,33 +1515,14 @@ int windNinjaCLI(int argc, char* argv[])
                 if(vm["write_wx_station_csv"].as<bool>()==true) //If the user wants an interpolated CSV
                 {
                     CPLDebug("STATION_FETCH", "Writing wxStation csv for step #%d", i_);
-                    if(vm.count("output_path")){
-                        pointInitialization::writeStationOutFile(windsim.getWxStations(i_),
-                                                   vm["output_path"].as<std::string>(),
-                                                   vm["elevation_file"].as<std::string>(),
-                                                   true);
-                    }
-                    else{
-                        pointInitialization::writeStationOutFile(windsim.getWxStations(i_),
-                                                   "",
-                                                   vm["elevation_file"].as<std::string>(),
-                                                   true);
-                    }
+                    pointInitialization::writeStationOutFile(windsim.getWxStations(i_), output_path, *elevation_file, true);
                 }
                 if(vm["write_wx_station_kml"].as<bool>() == true) //If the user wants a KML of the stations
                 {
                     CPLDebug("STATION_FETCH", "Writing wxStation kml for step #%d", i_);
-                    if(vm.count("output_path")){
-                        wxStation::writeKmlFile(windsim.getWxStations( i_ ),
-                                                vm["elevation_file"].as<std::string>(),
-                                                vm["output_path"].as<std::string>(), velocityUnits::getUnit(vm["output_speed_units"].as<std::string>()));
-                    }
-                    else
-                    {
-                        wxStation::writeKmlFile(windsim.getWxStations( i_ ),
-                                                vm["elevation_file"].as<std::string>(),
-                                                    "", velocityUnits::getUnit(vm["output_speed_units"].as<std::string>()));
-                    }
+                    wxStation::writeKmlFile(windsim.getWxStations( i_ ),
+                                                *elevation_file,
+                                                output_path, velocityUnits::getUnit(vm["output_speed_units"].as<std::string>()));
                 }
 
                 windsim.setOutputWindHeight( i_, vm["output_wind_height"].as<double>(),
@@ -1660,10 +1683,10 @@ int windNinjaCLI(int argc, char* argv[])
 
             //check if lcp to determine if surface veg needs to be set or not
             bool isLcp;
-            GDALDataset *poDS = (GDALDataset*)GDALOpen(vm["elevation_file"].as<std::string>().c_str(),
+            GDALDataset *poDS = (GDALDataset*)GDALOpen(elevation_file->c_str(),
                     GA_ReadOnly);
             if(poDS == NULL) {
-                fprintf(stderr, "Cannot open %s for reading, exiting...", vm["elevation_file"].as<std::string>().c_str());
+                fprintf(stderr, "Cannot open %s for reading, exiting...", elevation_file->c_str());
                 return -1;
             }
             const char *pszWkt = poDS->GetProjectionRef();
@@ -1677,6 +1700,8 @@ int windNinjaCLI(int argc, char* argv[])
             oSRS.importFromWkt((char**)&pszWkt);
             if(oSRS.IsGeographic())
             {
+                // PCM - try to convert to UTM
+
                 cerr << "Invalid DEM spatial reference, it is Geographic.\n";
                 return -1;
             }
@@ -1782,11 +1807,14 @@ int windNinjaCLI(int argc, char* argv[])
                         lengthUnits::getUnit(vm["units_shape_out_resolution"].as<std::string>()));
             }
             // those are all AAIGRID variants (in different projections and text formats)
-            if(option<bool>(vm,"write_ascii_output") || option<bool>(vm,"write_ascii_4326_output") || option<bool>(vm,"write_json_4326_output"))
+            if(option_val<bool>(vm,"write_ascii_output"))
             {
-                if (option<bool>(vm,"write_ascii_output")) windsim.setAsciiOutFlag( i_, true );
-                if (option<bool>(vm,"write_ascii_4326_output")) windsim.setAscii4326OutFlag( i_, true );
-                if (option<bool>(vm,"write_json_4326_output")) windsim.setJson4326OutFlag( i_, true );
+                windsim.setAsciiOutFlag( i_, true );
+                if (option_val<bool>(vm,"ascii_out_aaigrid")) windsim.setAsciiAaigridOutFlag( i_, true );
+                if (option_val<bool>(vm,"ascii_out_json")) windsim.setAsciiJsonOutFlag( i_, true );
+                if (option_val<bool>(vm,"ascii_out_4326")) windsim.setAscii4326OutFlag( i_, true );
+                if (option_val<bool>(vm,"ascii_out_utm")) windsim.setAsciiUtmOutFlag( i_, true );
+                if (option_val<bool>(vm,"ascii_out_uv")) windsim.setAsciiUvOutFlag( i_, true );
 
                 option_dependency(vm, "ascii_out_resolution", "units_ascii_out_resolution");
                 windsim.setAsciiResolution( i_, vm["ascii_out_resolution"].as<double>(),
