+++ src/ninja/ninjaArmy.h
@@ -69,6 +69,10 @@ extern boost::local_time::tz_database globalTimeZoneDB;
  * */
 #define IF_VALID_INDEX( i, iterable ) \
    if( i >= 0 && i < iterable.size() )
+
+#define CHECK_VALID_INDEX(i,iterable) \
+  if( i < 0 || i >= iterable.size() ) throw std::runtime_error("invalid index");
+
 /* *
  * Macro IF_VALID_INDEX_DO is a boiler plate for most of the ninjaArmy functions.
  * First, a range check is done on iterable to determine if 'i' is a valid index.
@@ -1122,7 +1126,34 @@ public:
     int setAsciiOutFlag( const int nIndex, const bool flag, char ** papszOptions=NULL );
 
     /**
-    * \brief Enable/disable AAIGRID ASCII output in EPSG:4326 projection for a ninja
+    * \brief Enable/disable AAIGRID (*.asc) output for a ninja
+    *
+    * \param nIndex index of a ninja
+    * \param flag   enable if true, disable if false
+    * \return errval Returns NINJA_SUCCESS upon success
+    */
+    int setAsciiAaigridOutFlag( const int nIndex, const bool flag, char ** papszOptions=NULL );
+    
+    /**
+    * \brief Enable/disable JSON output for a ninja
+    *
+    * \param nIndex index of a ninja
+    * \param flag   enable if true, disable if false
+    * \return errval Returns NINJA_SUCCESS upon success
+    */
+    int setAsciiJsonOutFlag( const int nIndex, const bool flag, char ** papszOptions=NULL );
+
+    /**
+    * \brief Enable/disable output in UTM projection for a ninja
+    *
+    * \param nIndex index of a ninja
+    * \param flag   enable if true, disable if false
+    * \return errval Returns NINJA_SUCCESS upon success
+    */
+    int setAsciiUtmOutFlag( const int nIndex, const bool flag, char ** papszOptions=NULL );
+
+    /**
+    * \brief Enable/disable output in EPSG:4326 projection for a ninja
     *
     * \param nIndex index of a ninja
     * \param flag   enable if true, disable if false
@@ -1131,13 +1162,13 @@ public:
     int setAscii4326OutFlag( const int nIndex, const bool flag, char ** papszOptions=NULL );
 
     /**
-    * \brief Enable/disable AAIGRID JSON output in EPSG:4326 projection for a ninja
+    * \brief Enable/disable output of u,v wind fields for a ninja
     *
     * \param nIndex index of a ninja
     * \param flag   enable if true, disable if false
     * \return errval Returns NINJA_SUCCESS upon success
     */
-    int setJson4326OutFlag( const int nIndex, const bool flag, char ** papszOptions=NULL );
+    int setAsciiUvOutFlag( const int nIndex, const bool flag, char ** papszOptions=NULL );
 
     /**
     * \brief Set the resoultion of ASCII output for a ninja
