+++ src/ninja/windninja.cpp
@@ -206,7 +206,7 @@ WINDNINJADLL_EXPORT NinjaErr NinjaStartRuns
         }
         catch( ... )
         {
-            handleException();
+            return handleException();
         }
     }
     else
@@ -749,8 +749,7 @@ WINDNINJADLL_EXPORT const char * NinjaGetInitializationMethod
 {
     if( NULL != ninja )
     {
-        return reinterpret_cast<ninjaArmy*>( ninja )->getInitializationMethodString
-            ( nIndex ).c_str();
+        return strdup(reinterpret_cast<ninjaArmy*>(ninja)->getInitializationMethodString(nIndex).c_str());
     }
     else
     {
@@ -952,9 +951,10 @@ WINDNINJADLL_EXPORT NinjaErr  NinjaSetNumVertLayers
 WINDNINJADLL_EXPORT NinjaErr NinjaSetOutputPath
     ( NinjaH * ninja, const int nIndex, const char * path)
 {
-    if( NULL != ninja )
-    {
+    if( NULL != ninja ){
         return reinterpret_cast<ninjaArmy*>( ninja )->setOutputPath( nIndex, std::string( path ) );
+    } else {
+        return NINJA_E_NULL_PTR;
     }
 }
 
@@ -974,9 +974,10 @@ WINDNINJADLL_EXPORT NinjaErr NinjaSetOutputPath
 WINDNINJADLL_EXPORT const double* NinjaGetOutputSpeedGrid
     ( NinjaH * ninja, const int nIndex )
 {
-    if( NULL != ninja )
-    {
+    if( NULL != ninja ) {
            return reinterpret_cast<ninjaArmy*>( ninja )->getOutputSpeedGrid( nIndex );
+    } else {
+        return NULL;
     }
 }
 
@@ -996,9 +997,10 @@ WINDNINJADLL_EXPORT const double* NinjaGetOutputSpeedGrid
 WINDNINJADLL_EXPORT const double* NinjaGetOutputDirectionGrid
     ( NinjaH * ninja, const int nIndex )
 {
-    if( NULL != ninja )
-    {
-           return reinterpret_cast<ninjaArmy*>( ninja )->getOutputDirectionGrid( nIndex );
+    if( NULL != ninja ) {
+        return reinterpret_cast<ninjaArmy*>( ninja )->getOutputDirectionGrid( nIndex );
+    } else {
+        return NULL;
     }
 }
 
@@ -1019,9 +1021,10 @@ WINDNINJADLL_EXPORT const double* NinjaGetOutputDirectionGrid
 WINDNINJADLL_EXPORT const char* NinjaGetOutputGridProjection
     ( NinjaH * ninja, const int nIndex )
 {
-    if( NULL != ninja )
-    {
-           return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridProjection( nIndex );
+    if( NULL != ninja ) {
+        return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridProjection( nIndex );
+    } else {
+        return NULL;
     }
 }
 
@@ -1042,9 +1045,10 @@ WINDNINJADLL_EXPORT const char* NinjaGetOutputGridProjection
 WINDNINJADLL_EXPORT const double NinjaGetOutputGridCellSize
     ( NinjaH * ninja, const int nIndex )
 {
-    if( NULL != ninja )
-    {
-           return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridCellSize( nIndex );
+    if( NULL != ninja ) {
+        return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridCellSize( nIndex );
+    } else {
+        throw std::runtime_error("no ninjaArmy");
     }
 }
 
@@ -1065,9 +1069,10 @@ WINDNINJADLL_EXPORT const double NinjaGetOutputGridCellSize
 WINDNINJADLL_EXPORT const double NinjaGetOutputGridxllCorner
     ( NinjaH * ninja, const int nIndex )
 {
-    if( NULL != ninja )
-    {
-           return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridxllCorner( nIndex );
+    if( NULL != ninja ) {
+        return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridxllCorner( nIndex );
+    } else {
+        throw std::runtime_error("no ninjaArmy");
     }
 }
 
@@ -1088,9 +1093,10 @@ WINDNINJADLL_EXPORT const double NinjaGetOutputGridxllCorner
 WINDNINJADLL_EXPORT const double NinjaGetOutputGridyllCorner
     ( NinjaH * ninja, const int nIndex )
 {
-    if( NULL != ninja )
-    {
-           return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridyllCorner( nIndex );
+    if( NULL != ninja ) {
+        return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridyllCorner( nIndex );
+    } else {
+        throw std::runtime_error("no ninjaArmy");
     }
 }
 
@@ -1111,9 +1117,10 @@ WINDNINJADLL_EXPORT const double NinjaGetOutputGridyllCorner
 WINDNINJADLL_EXPORT const int NinjaGetOutputGridnCols
     ( NinjaH * ninja, const int nIndex )
 {
-    if( NULL != ninja )
-    {
-           return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridnCols( nIndex );
+    if( NULL != ninja ) {
+        return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridnCols( nIndex );
+    } else {
+        throw std::runtime_error("no ninjaArmy");
     }
 }
 
@@ -1134,9 +1141,10 @@ WINDNINJADLL_EXPORT const int NinjaGetOutputGridnCols
 WINDNINJADLL_EXPORT const int NinjaGetOutputGridnRows
     ( NinjaH * ninja, const int nIndex )
 {
-    if( NULL != ninja )
-    {
-           return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridnRows( nIndex );
+    if( NULL != ninja ) {
+        return reinterpret_cast<ninjaArmy*>( ninja )->getOutputGridnRows( nIndex );
+    } else {
+        throw std::runtime_error("no ninjaArmy");
     }
 }
 
@@ -1468,9 +1476,10 @@ WINDNINJADLL_EXPORT NinjaErr NinjaSetTxtOutFlag
 WINDNINJADLL_EXPORT const char * NinjaGetOutputPath
     ( NinjaH * ninja, const int nIndex )
 {
-    if( NULL != ninja )
-    {
-        return reinterpret_cast<ninjaArmy*>( ninja )->getOutputPath( nIndex ).c_str();
+    if( NULL != ninja ) {
+        return strdup(reinterpret_cast<ninjaArmy*>( ninja )->getOutputPath( nIndex ).c_str());
+    } else {
+        return NULL;
     }
 }
